requirements.txt
pytorch-lightning == 1.7.0
torchvision == 0.13.1
hydra-core == 1.2.0